refactor into intake file
